Delete old App.java and move to App.kt
